% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.scelestial}
\alias{.scelestial}
\title{Internal function for running scelestial algorithm.}
\usage{
.scelestial(data, minK = 3L, maxK = 4L)
}
\arguments{
\item{data}{The data}

\item{minK, maxK}{Minimum and maximum number of vertices to be considered 
for k-restricted Steiner tree.}
}
\value{
The tree as well as missing value imputation
}
\description{
Internal function for running scelestial algorithm.
}
