% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Li-data.R
\docType{data}
\name{Li}
\alias{Li}
\title{Bladder invasive single cell tumor dataset}
\format{
Each column represent a cell and each row represent a locus. 
"./." represent the missing value, "A/A" the normal state and "C/C" the mutated state.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3626503/bin/2047-217X-1-12-S2.xls}{QTL Archive}
}
\usage{
data(Li)
}
\description{
Bladder invasive single cell tumor dataset
}
\examples{
data(Li)
}
\references{
Gigascience. 2012 Aug 14;1(1):12. doi: 10.1186/2047-217X-1-12.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23587365}{PubMed})
}
\keyword{datasets}
