% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scelestial.R
\name{my.dfs}
\alias{my.dfs}
\title{Runs DFS on tree and calculates parent of each node
as well as depth and upper-depth of nodes.}
\usage{
my.dfs(graph, root = NULL)
}
\arguments{
\item{graph}{The tree}

\item{root}{The starting node of DFS.}
}
\value{
a list with \code{father} representing the parent node, and
  \code{balance.depth} representing the distance between the node and
  the farthest node to it, as the elements.
}
\description{
It is used for internal purposes.
}
