% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scelestial.R
\name{my.general.dfs}
\alias{my.general.dfs}
\title{Running depth first search on a tree and calling
functions on entrance/exit events}
\usage{
my.general.dfs(
  nei,
  v,
  f,
  extra,
  in.call,
  mid.call.before,
  mid.call.after,
  out.call
)
}
\arguments{
\item{nei}{Neighbor list for each vertex}

\item{v}{Starting node}

\item{f}{Parent node}

\item{extra}{the shared object for the whole DFS}

\item{in.call}{First function to call}

\item{mid.call.before}{Function to call before calling child DFS}

\item{mid.call.after}{Function to call after calling child DFS}

\item{out.call}{Last function to call}
}
\value{
the \code{extra} parameter modified with \code{in.call}, \code{mid.call.before}, \code{mid.call.after}, and \code{out.call} functions
}
\description{
It is used for internal purposes.
}
