% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scelestial.R
\name{read.sequence.table}
\alias{read.sequence.table}
\title{Read mutation table}
\usage{
read.sequence.table(file.name)
}
\arguments{
\item{file.name}{Name of the file to be loaded}
}
\value{
A table representing the content of the file. 
  First column of the file represents the row names.
}
\description{
A simple read of a sequencing file.
}
\examples{
# An example input without header could be like following:
# 1 C/C A/A A/A A/A 
# 2 ./. A/A C/C C/C 
# 3 C/C A/A C/C ./. 
# 4 A/A ./. ./. ./. 
# 5 ./. A/A A/A A/A 
# 
# For this file you can run
read.sequence.table(system.file("extdata/sample1.txt", package="RScelestial"))

}
