% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RScelestial-package.R
\docType{package}
\name{RScelestial}
\alias{RScelestial}
\title{RScelestial: An R wrapper for scelestial algorithm for single-cell lineage tree reconstruction
  through an approximation algorithm based on Steiner tree problem}
\description{
This package provides a wrapper for the scelestial which is implemented in C++.
The package contains function \code{scelestial} for running the algorithm and 
\code{synthesis} for tumor simulation for providing synthetic data.
}
