% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scelestial.R
\name{tree.plot}
\alias{tree.plot}
\title{Plotting the tree}
\usage{
tree.plot(graph, ...)
}
\arguments{
\item{graph}{Output of scelestial or the G element of the scelestial output.}

\item{...}{Parameters passing to the plot function}
}
\description{
Plotting the igraph tree created by scelestial.
}
